Set HomeView and ProfileView + ProfileConnection (To debug, reload from Internal Storage)

+++ AndroidStudioProjects/ProjectMM/app/src/main/java/com/example/projectmm/ScanQRActivity.kt
@@ -37,9 +37,8 @@ class ScanQRActivity : AppCompatActivity(), ZXingScannerView.ResultHandler {
         Log.d("QRCode2", rawResult.barcodeFormat.toString())
 
         val intent = Intent(this, MovieDetailsActivity::class.java)
-        intent.putExtra("movie_id", rawResult.text)
+        intent.putExtra("movie_id", rawResult.text.toInt())
         startActivity(intent)
-
 
         // Note:
         // * Wait 2 seconds to resume the preview.
